+++ packages/admin/admin/src/stack/breadcrumbs/StackBreadcrumbs.styles.tsx
@@ -1,39 +1,64 @@
-import { BreadcrumbsClassKey } from "@mui/material";
 import { Theme } from "@mui/material/styles";
 import { createStyles } from "@mui/styles";
 
 import { StackBreadcrumbsProps } from "./StackBreadcrumbs";
 
-export type StackBreadcrumbsClassKey = BreadcrumbsClassKey | "link" | "last";
+export type StackBreadcrumbsClassKey =
+    | "root"
+    | "breadcrumbs"
+    | "listItem"
+    | "link"
+    | "disabledLink"
+    | "overflowLink"
+    | "separator"
+    | "backButton"
+    | "backButtonSeparator";
 
-export const styles = ({ palette }: Theme) => {
-    return createStyles<StackBreadcrumbsClassKey, StackBreadcrumbsProps>({
+export const styles = ({ palette, typography }: Theme) =>
+    createStyles<StackBreadcrumbsClassKey, StackBreadcrumbsProps>({
         root: {
-            paddingTop: 30,
-            paddingBottom: 30,
+            position: "relative",
         },
-        ol: {},
-        li: {},
-        separator: {
-            color: palette.grey[300],
-            "& [class*='MuiSvgIcon-root']": {
-                fontSize: 12,
-            },
-            link: {
-                color: palette.text.primary,
-                textDecoration: "underline",
-                "& [class*='MuiTypography']": {
-                    fontSize: 13,
-                    lineHeight: "14px",
-                },
-            },
+        breadcrumbs: {
+            display: "flex",
+            height: 50,
+            borderBottom: `1px solid ${palette.divider}`,
+            boxSizing: "border-box",
+            flexWrap: "nowrap",
+            overflowX: "auto", // Make the breadcrumbs scrollable, if they still take up too much space, when only the first, last & the overflow link are visible.
         },
-        link: {},
-        last: {
+        listItem: {
+            display: "flex",
+            alignItems: "center",
+            flexShrink: 0,
+            whiteSpace: "nowrap",
+        },
+        link: {
+            fontSize: 13,
+            lineHeight: "14px",
+            fontWeight: typography.fontWeightMedium,
+            color: palette.grey[600],
+            textDecorationColor: "currentColor",
+        },
+        disabledLink: {
             color: palette.text.disabled,
-            "&, &:hover": {
-                textDecoration: "none",
-            },
+        },
+        overflowLink: {
+            cursor: "pointer",
+            paddingTop: 12,
+            paddingBottom: 12,
+        },
+        separator: {
+            fontSize: 12,
+            lineHeight: 0,
+            marginLeft: 8,
+            marginRight: 8,
+        },
+        backButton: {},
+        backButtonSeparator: {
+            height: 30,
+            width: 1,
+            backgroundColor: palette.divider,
+            marginRight: 12,
         },
     });
-};